+++ R/rcpp.R
@@ -1,9 +1,11 @@
-#' Use Rcpp
+#' Use C and C++.
 #'
-#' Creates `src/` and adds needed packages to `DESCRIPTION`.
+#' Creates `src/` and adds needed packages to `DESCRIPTION`, and
+#' optionally creates `.c` or `.cpp` file.
 #'
+#' @param name If supplied, creates and opens `src/name.{c,cpp}`.
 #' @export
-use_rcpp <- function() {
+use_rcpp <- function(name = NULL) {
   check_is_package("use_rcpp()")
   check_uses_roxygen("use_rcpp()")
 
@@ -13,7 +15,29 @@ use_rcpp <- function() {
   use_dependency("Rcpp", "Imports")
   roxygen_ns_append("@importFrom Rcpp sourceCpp") && roxygen_update()
 
+  if (!is.null(name)) {
+    name <- slug(name, "cpp")
+    check_file_name(name)
+
+    use_template("code.cpp", path("src", name), open = TRUE)
+  }
+
   invisible()
+}
+
+#' @rdname use_rcpp
+#' @export
+use_c <- function(name = NULL) {
+  use_src()
+
+  if (!is.null(name)) {
+    name <- slug(name, "c")
+    check_file_name(name)
+
+    use_template("code.c", path("src", name), open = TRUE)
+  }
+
+  invisible(TRUE)
 }
 
 use_src <- function() {